+++ bin/gui/windows/popup/__init__.py
@@ -1,4 +1,4 @@
-from .base_popup import BasePopUp
-from .classic_popup import PopUp
-from .warning_popup import WarningPopUp
-from .yes_no_popup import YesNoPopUp
+from .base import BasePopUp
+from .default import PopUp
+from .warning import WarningPopUp
+from .yes_no import YesNoPopUp
