+++ bin/gui/windows/popup/yes_no.py
@@ -3,7 +3,7 @@ from typing import Callable
 from PySide6.QtCore import Qt
 from PySide6.QtWidgets import QWidget
 
-from .base_popup import BasePopUp
+from .base import BasePopUp
 from bin.gui.widgets.buttons.popup import PopUpDefaultButton
 from bin.gui.widgets.labels.popup import PopUpLabel
 from bin.gui.widgets.layouts import HorizontalLayout
@@ -13,10 +13,10 @@ class YesNoPopUp(BasePopUp):
     def __init__(
         self,
         parent: QWidget,
-        title: str or None,
         message_text: str,
-        on_click_accept: Callable or None,
-        on_click_reject: Callable or None,
+        title: str or None = None,
+        on_click_accept: Callable or None = None,
+        on_click_reject: Callable or None = None,
     ) -> None:
         self.message_text = message_text
         self.on_click_accept = self._decorator_onclick(on_click_accept)
