--- src/styles/components/windows/popup/image_label_popup/__init__.py
@@ -1,3 +1,0 @@
-from .object import (
-    DEFAULT_IMAGE_LABEL_POP_UP_STYLESHEET,
-)
