--- src/styles/components/windows/popup/image_label_popup/object.py
@@ -1,4 +1,0 @@
-from src.styles.stylesheet import stylesheet
-
-
-DEFAULT_IMAGE_LABEL_POP_UP_STYLESHEET = stylesheet(__file__)
